+++ src/core/plugins/MidaPluginParameters.ts
@@ -2,6 +2,7 @@ import { MidaPluginActions } from "#plugins/MidaPluginActions";
 import { GenericObject } from "#utilities/GenericObject";
 
 export type MidaPluginParameters = {
+    id: string;
     name: string;
     description?: string;
     version: string;