+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* Create "evaluateOnNewDocument" method for browser tabs.
+
 1.0.0 - 09-05-2021
 ===================