+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Expose puppeteer page instance in browser tab.
 * Define monthly timeframe type.
 * Create "rightClick" method for browser tabs.
 * Create "evaluateOnNewDocument" method for browser tabs.