+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Remove right click method from browser tabs and create options object in the click method.
 * Return string symbols instead of objects when requesting the account symbols.
 * Remove export of spread type.
 * Remove spread type property from symbols.