+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Now the broker accounts "getSymbolLastTick" method can return undefined.
+* Create "getSymbolBid" and "getSymbolAsk" methods for broker accounts.
 * Create "tick" event for broker orders.
 * Create "initiator" field for broker orders.
 * Extend error types enumeration.