+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+2.0.0 - 28-05-2021
+===================
 * Create "removeEventListener" method for broker accounts and broker orders.
 * Create new logo.
 * Now the broker accounts "getSymbolLastTick" method can return undefined.